+++ src/vtex/index.ts
@@ -2,10 +2,12 @@ import MasterData from './MasterData';
 import getProductById from './getProductById';
 import getProductSuggestions from './getProductSuggestions';
 import getProductVariations from './getProductVariations';
+import getCollectionShelf from './getCollectionShelf';
 
 export {
   MasterData,
   getProductById,
   getProductSuggestions,
   getProductVariations,
+  getCollectionShelf,
 };