+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [utils] Added `isAndroid` method.
 - [vtex] Added `getProductSuggestions` method.
 - [vtex] Added `getProductVariations` method.
+- [vtex] Added `getCollectionShelf` method.
 - [vtex] Added `MasterData.put` method.
 - [vtex] Added `MasterData.patch` method.